Attempt to fix slideshare
