+++ core/routes/youtube.py
@@ -1,17 +1,9 @@
-from core import app, socketio
+from core import app, file_data, status
 from flask import render_template, send_file, request, session, flash, url_for, redirect
 from pytube import YouTube, Playlist
 import pytube.exceptions as exceptions
-from decouple import config
 from core.utils import playlist
 from core import yt
-from flask_socketio import send
-from threading import Thread
-from time import sleep
-
-
-file_data = {}
-status = {}
 
 
 @app.route('/yt-downloader/video', methods=['GET', 'POST'])
@@ -51,26 +43,6 @@ def start_preparation(msg, url, itag):
     file_data.update(fname=filename)
     file_data.update(status="Done")
     status.update({f"{msg}": "Download-Ready"})
-
-
-@socketio.on('message')
-def socket_bidirct(msg):
-
-    if msg[0] != "User has connected!":
-        url = session['video_link']
-        t = Thread(target=start_preparation, args=(
-            msg[0], url, msg[1],), daemon=True)
-        t.start()
-
-        while True:
-            sleep(2)
-            if status.get(msg[0]) == "Download-Ready":
-                send("Download-Ready")
-                break
-        del status[msg[0]]
-
-    if msg[0] == "User has connected!":
-        print(msg[0])
 
 
 @app.post('/yt-downloader/video/download')
